requirements.txt
easydict
terminaltables
torchsummary